refactor: split builder API stubs by feature
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
