+++ src/features/runs/api/index.ts
@@ -1,0 +1,7 @@
+import { API_BASE } from "@/shared/config/env";
+import type { BuildRun, BuildSpec } from "@/shared/lib/types";
+
+export async function executeBuild(_spec: BuildSpec): Promise<BuildRun> {
+  void API_BASE;
+  throw new Error("Not implemented");
+}
