+++ src/shared/config/env.ts
@@ -1,0 +1,2 @@
+export const API_BASE =
+  import.meta.env.VITE_BUILDER_API_URL ?? "http://localhost:8000";
